news articles crd

+++ app/Http/Controllers/TournamentController.php
@@ -68,21 +68,38 @@ class TournamentController extends Controller
       $tournament->update();
 
       if ($request->has('image')) {
-        $image = Image::find($tournament->image_id);
+        if (!$tournament->image_id) {
+          $file = $request->file('image');
+          $extension = $file->extension();
+          $fileName = $tournament->slug . '.' . $extension;
+          $file->move(public_path('/images/tournaments/'), $fileName);
 
-        file_exists(public_path($image->src))
-          && unlink(public_path($image->src));
+          $image = Image::create([
+            'title' => $fileName,
+            'src' => '/images/tournaments/' . $fileName,
+            'alt' => $tournament->title,
+            'extension' => $extension,
+          ]);
 
-        $file = $request->file('image');
-        $extension = $file->extension();
-        $fileName = $tournament->slug . '.' . $extension;
-        $file->move(public_path('/images/tournaments/'), $fileName);
+          $tournament->image_id = $image->id;
+          $tournament->update();
+        } else {
+          $image = Image::find($tournament->image_id);
 
-        $image->title = $fileName;
-        $image->src = '/images/tournaments/' . $fileName;
-        $image->alt = $tournament->title;
-        $image->extension = $extension;
-        $image->update();
+          file_exists(public_path($image->src))
+            && unlink(public_path($image->src));
+
+          $file = $request->file('image');
+          $extension = $file->extension();
+          $fileName = $tournament->slug . '.' . $extension;
+          $file->move(public_path('/images/tournaments/'), $fileName);
+
+          $image->title = $fileName;
+          $image->src = '/images/tournaments/' . $fileName;
+          $image->alt = $tournament->title;
+          $image->extension = $extension;
+          $image->update();
+        }
       }
 
       return Tournament::with('image')->find($id);
